@@ -1,10 +1,10 @@
 package uk.gov.justice.digital.hmpps.sendlegalmailtoprisonsapi.barcode
 
+import jakarta.persistence.Entity
+import jakarta.persistence.Id
+import jakarta.persistence.Table
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import javax.persistence.Entity
-import javax.persistence.Id
-import javax.persistence.Table
 
 @Repository
 interface BarcodeRepository : JpaRepository<Barcode, String>
