@@ -1,14 +1,17 @@
 package uk.gov.justice.digital.hmpps.sendlegalmailtoprisonsapi.config
 
-import com.microsoft.applicationinsights.web.internal.ThreadContext
+import com.nimbusds.jwt.JWTClaimsSet
+import com.nimbusds.jwt.SignedJWT
+import io.opentelemetry.api.trace.Span
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
+import org.slf4j.LoggerFactory
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.servlet.HandlerInterceptor
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-import uk.gov.justice.digital.hmpps.sendlegalmailtoprisonsapi.security.JwtService
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import java.text.ParseException
 
 @Configuration
 @ConditionalOnExpression("T(org.apache.commons.lang3.StringUtils).isNotBlank('\${applicationinsights.connection.string:}')")
@@ -20,19 +23,37 @@ class BarcodeTokenClientTrackingConfiguration(private val barcodeTokenClientTrac
 }
 
 @Configuration
-class BarcodeTokenClientTrackingInterceptor(private val jwtService: JwtService) : HandlerInterceptor {
+class BarcodeTokenClientTrackingInterceptor : HandlerInterceptor {
+
+  companion object {
+    private val LOG = LoggerFactory.getLogger(this::class.java)
+  }
+
   override fun preHandle(request: HttpServletRequest, response: HttpServletResponse, handler: Any): Boolean {
-    val properties = ThreadContext.getRequestTelemetryContext().httpRequestTelemetry.properties
     val token = request.getHeader("create-barcode-token")
-    token
-      ?.let { jwtService.subject(it) }
-      ?.also { properties["username"] = it }
-    token
-      ?.let { jwtService.clientId(it) }
-      ?.also { properties["clientId"] = it }
-
-    properties["slm_Client_IP"] = request.getHeader("x-slm-client-ip") ?: ""
-
+    token?.let {
+      try {
+        val jwtBody = getClaimsFromJWT(token)
+        val user = jwtBody.getClaim("user_name")?.toString()
+        val currentSpan = getCurrentSpan()
+        user?.run {
+          currentSpan.setAttribute("username", this) // username in customDimensions
+        }
+        currentSpan.setAttribute("clientId", jwtBody.getClaim("client_id").toString())
+        currentSpan.setAttribute("slm_Client_IP", request.getHeader("x-slm-client-ip") ?: "")
+      } catch (e: ParseException) {
+        LOG.warn("problem decoding jwt public key for application insights", e)
+      }
+    }
     return true
+  }
+
+  fun getCurrentSpan(): Span {
+    return Span.current()
+  }
+
+  @Throws(ParseException::class)
+  private fun getClaimsFromJWT(token: String): JWTClaimsSet {
+    return SignedJWT.parse(token.replace("Bearer ", "")).jwtClaimsSet
   }
 }
