@@ -23,6 +23,6 @@ class OpenApiConfiguration(buildProperties: BuildProperties) {
       Info().title("Send Legal Mail to Prisons")
         .version(version)
         .description("Create and receive Rule 39 mail")
-        .contact(Contact().name("HMPPS Digital Studio").email("feedback@digital.justice.gov.uk")),
+        .contact(Contact().name("Visit Someone In Prison").email("prisonvisitsbooking@digital.justice.gov.uk")),
     )
 }
