@@ -1,12 +1,12 @@
 package uk.gov.justice.digital.hmpps.sendlegalmailtoprisonsapi.contact
 
+import jakarta.validation.Validation
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import java.time.LocalDate
-import javax.validation.Validation
 
 class ContactRequestTest {
 
