@@ -1,11 +1,10 @@
 package uk.gov.justice.digital.hmpps.sendlegalmailtoprisonsapi.config
 
-import com.microsoft.applicationinsights.web.internal.RequestTelemetryContext
-import com.microsoft.applicationinsights.web.internal.ThreadContext
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeEach
+import io.opentelemetry.api.common.AttributeKey
+import io.opentelemetry.api.trace.Tracer
+import io.opentelemetry.sdk.testing.junit5.OpenTelemetryExtension
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.RegisterExtension
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.mock.web.MockHttpServletRequest
 import org.springframework.mock.web.MockHttpServletResponse
@@ -16,15 +15,7 @@ class BarcodeTokenClientTrackingInterceptorTest : IntegrationTest() {
   @Autowired
   private lateinit var barcodeTokenClientTrackingInterceptor: BarcodeTokenClientTrackingInterceptor
 
-  @BeforeEach
-  fun setup() {
-    ThreadContext.setRequestTelemetryContext(RequestTelemetryContext(1L))
-  }
-
-  @AfterEach
-  fun tearDown() {
-    ThreadContext.remove()
-  }
+  private val tracer: Tracer = otelTesting.openTelemetry.getTracer("test")
 
   @Test
   fun `should add client and username to telemetry`() {
@@ -32,10 +23,21 @@ class BarcodeTokenClientTrackingInterceptorTest : IntegrationTest() {
     val req = MockHttpServletRequest()
     req.addHeader("create-barcode-token", token)
     val res = MockHttpServletResponse()
-    barcodeTokenClientTrackingInterceptor.preHandle(req, res, "null")
-    val insightTelemetry = ThreadContext.getRequestTelemetryContext().httpRequestTelemetry.properties
+    tracer.spanBuilder("span").startSpan().run {
+      makeCurrent().use { barcodeTokenClientTrackingInterceptor.preHandle(req, res, "null") }
+      end()
+    }
+    otelTesting.assertTraces().hasTracesSatisfyingExactly({ t ->
+      t.hasSpansSatisfyingExactly({
+        it.hasAttribute(AttributeKey.stringKey("username"), "some.email@company.com")
+        it.hasAttribute(AttributeKey.stringKey("clientId"), "send-legal-mail")
+      },)
+    },)
+  }
 
-    assertThat(insightTelemetry["username"]).isEqualTo("some.email@company.com")
-    assertThat(insightTelemetry["clientId"]).isEqualTo("send-legal-mail")
+  private companion object {
+    @JvmStatic
+    @RegisterExtension
+    private val otelTesting: OpenTelemetryExtension = OpenTelemetryExtension.create()
   }
 }
